Bypass Dell BIOS config pre checks to be skipped with env var
The runtime pre-requisite checks are specific to CentOS environments,
and we want to be able to run this on other Linux OSes. By setting the
UTIL_RACADM7 env var to the path of your racadm binary, ironlib will
skip the pre() check and use the specified path to run racadm.