Add release workflows for ironlib container
This adds a couple of workflows to build the ironlib container to the
GitHub Container Registry. This replaces `tag-publish.yaml` and instead
uses a common workflow to do the build [1].

There will be a `latest` tag that gets updated with every push to the
`main` branch, as well as release-specific container tags for each git
tag pushed with the release format `v**`.

[1] https://github.com/metal-toolbox/container-push

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>