common.Device: replace model.Device with common.Device (#48)
In the previous set of changes - 408ed07, the common.Device{} object
was embedded in model.Device{} to continue including the OemComponents field
for OEM hardware. This change moves the OemComponents field into
the Hardware{} object and so the model.Device{} object is removed
and ironlib will now return a common.Device{} for consistency with bmclib.