Update .golangci.yml from metal-toolbox/golangci-lint-config
And fix the errors of course, gotta keep CI green! Most of the class of
changes are pretty small.

The big one here is from contextcheck which definitely is correct that
we should pass context.Context from calling functions down to called
functions when the called function is using a context.Context. The
question is are we *sure* we want to tie those function calls deep down
with the top level context.Context? These all seem like the answer is
yes which is why I did them, but will need to make sure reviewers are
aware and take a good look.