utils/nvme: Use nvme cli internal names for capabilities
These dynamic/generated names don't really work as can probably be
noticed by adding the `fna` and `sanicap` synthetic capabilities. They
don't work because nvme gives different output for enabled/disable,
supported/unsupported, all/single capabilities. So we end up with
enabled/disabled is detected by presence vs absence and we have to look
up two "different" capabilities to determine the status of an _actual_
capability.

So instead of doing that dynamic name we just use the internal nvme
identifiers seen in `nvme-print-stdout.c` at this point in time. We will
not change them later if nvme changes them internally. This way we end
up being more likely to be future proof. It would be much easier for
nvme-cli to tweak the string returned in the human readable output than
it would be for the linux kernel to change meaning of bits returned in
the ioctl call.