Get rid of nolint:gocritic comments
2 ignores needed some small code changes to make it possible to avoid
commenting gocritic out. Both in tests and both solved by using stdlib
helpers.

I replaced a os.MkdirTemp() with t.TempDir() that gets cleaned up when
the test is done, this avoids needing to clean up the sub-dirs. I could
have cleaned up tmpDir instead of using t.TempDir for the same effect
but its better to let the stdlib do it if it wants to.

The other change was to run loop tests under t.Run() so that gocritic
stops complaining about defer-in-a-loop cases. This was probably the
better option from the beginning anyway since it lets us test
individually failing cases easily.