Be consistent with setting/using trace
Not a fan of how this ended up but at least its all consistent. Before
this change the code was all over the place with how it decided to check
if Trace was enabled or not (== vs >=). Also most structs had a trace
field that was never set and thus not needed but I decided to use the
field instead of removing it since it is being used by the firmware
checksum collector and I'd rather they all do it the same way.