utils/watermark: include watermarksSize in offset limit
The previous code was subtly wrong and only caught by random chance when
I ran some tests. We need to limit the random offset so that we can be
sure to write the full buffer without crossing over into the next chunk,
but this needs to happen when getting the random offset and not
calculating the position.

Consider the following situation:

* Disk size is 5120B (chunks are 512B)
* First iteration of the loop so chunkStart = 0
* rand.Int randomly returns 1
* randomPosition is calculated as (0 + 1 - 512) == -511
* We try to seek to position -511 of the disk
* Error

Where now it looks like:

* Disk size is 5120B (chunks are 512B)
* First iteration of the loop so chunkStart = 0
* rand.Int randomly returns 1
* randomPosition is calculated as (0 + 1) == 1
* We try to seek to position 1 of the disk
* Ok

Likewise we are still save at the other extreme:

* Disk size is 5120B (chunks are 512B)
* First iteration of the loop so chunkStart = 0
* rand.Int will only ever return 0 since
  chunkSize-watermarksSize=512-512=0
* randomPosition is calculated as (0 + 0) == 0
* We try to seek to position 0 of the disk
* Ok

Or at the end of the disk:

* Disk size is 5120B (chunks are 512B)
* Last iteration of the loop so chunkStart = 4608
* rand.Int will only ever return 0 since
  chunkSize-watermarksSize=512-512=0
* randomPosition is calculated as (4608 + 0) == 4608
* We try to seek to position 4608 of the disk
* Ok
* Write 512B
* Cursor is now at 5120, right at the end of the disk and Ok