utils/watermark: Refactor tests
A few simple refactors to make the tests easier to reason about. Instead
of sharing the testfile and wiping it we now use a small function and
t.TempDir that creates the testfile per test and cleans up after exit.
No need to wipe the test file at the beginning of each test.

Used more assertions from stretchr/testify instead of `if` checks, they
just read better.

Added a test for unsuccessful wipe.

Renamed the subtests to be more descriptive.