utils/watermark: Ensure file is large enough for watermarking
The code was not quite right before, even though there was a test that
seemed to be checking this. The test actually failed when seeking past
the end of the file. Some might say same thing but it breaks other code
and test that I was writing that fixes something else in here.