utils/watermark: Fix for loop terminating condition in writeWatermarks
The terminating condition was previously wrong if fileSize was not a
multiple of watermarksCount. In this case we'd have some left over bytes
in an 11th chunk and the chunkStart < fileSize is true for the next
iteration but seeking would fail. So lets iterate over the
watermarksCount count instead.

For example, here's a run where I came across this on a real nvme device
(coming from the future ;D ):
```
chunkStart: 230454462870 fileSize: 256060514304
   seek-to: 234037338277      end: 234037338789
chunkStart: 256060514300 fileSize: 256060514304
   seek-to: 280780001714      end: 280780002226
```
The final seek is past the end of the device and returns an error.