dmidecode: Handle binary lookup like other utilities
The github.com/dselans/dmidecode package has a static list of paths to
look for the dmidecode binary when used with dmidecode.New, but we don't
really do things like that here. For example maybe we want to pass in a
different name/path for the binary via env var like we do for other
tools. So this commit avoids dmidecode.New so that it can follow along
with the pattern already in-use here.