model: Get rid of OemComponents
This seems like the wrong layout for this functionality. Are we going to
have to add a new member for every manufacturer that has OEM specific
components? It doesn't really sound like a good idea. Instead we can
lean on Hardware.Device.Vendor to have dell/$manufacturer and then any
OEMComponents are _obviously_ for Dell/$manufacturer.