SQUASHED, NEED TO BREAK UP BETTER
--- Define and use our own Drive type

Going to invert how drive wiping works by having the drive know how to
wipe itself, so we need our own type to be able to give it methods.

---

Make Nvme a WipersGetter and pass to drive.New

Now we have a way to get the wipe utitlies for an nvme drive easily.