Rename Disk, device -> Drive, logicalName
Drive is already a well known/used name for disk in this repo and
bmc-toolbox so lets go with that instead of mixing Drive and Disk. Its
unfortunate that we have Disk in RAID stuff but thats also an issue in
bmc-toolbox/common. Changing would need to wait for major releases.

Same deal with device -> drive, device looks like it would be a
common.Device which is something totally different.