Check file contents in Test_FillZeroWipeDrive
Previous test wasn't really testing the file being filled with zeros,
just that it doesn't change size. Now we write random data to the file,
verify the contents to *not* be just zeros, proceed with writing zeros
and the finally verify the new contents is just zeros. Along the way
make sure the file also hasn't changed size.