Make WipeDrive verify the wipe
This is in line with what Nvme.WipeDrive already does. Nvme has to do it
that way because it needs to verify the method tried actually worked or
fallback to another one. No point in returning no-error to the caller if
the disk wasn't actually wiped.

Had to bump up the file sizes in Test_FillZeroWipeDrive because
ApplyWatermarks needs more than 4KiB worth of space to write all the
watermarks (512 * 10 == 5120).